Task2/requirements.txt
streamlit
torch
diffusers
google-generativeai
xformers
accelerate
transformers
sentence-transformers